vintner setup clean --force

+++ src/controller/setup/clean.ts
@@ -1,9 +1,20 @@
 import config from '#config'
 import * as files from '#files'
+import {Instances} from '#repository/instances'
+import {Templates} from '#repository/templates'
 import * as os from 'os'
 import path from 'path'
 
-export default async function () {
+export type SetupCleanOptions = {force: Boolean}
+
+export default async function (options: SetupCleanOptions) {
+    // Ensure no templates or instances exist
+    options.force = options.force ?? false
+    if (!options.force) {
+        if (!Templates.isEmpty()) throw new Error(`Templates not empty`)
+        if (!Instances.isEmpty()) throw new Error(`Instances not empty`)
+    }
+
     // Delete home directory
     files.deleteDirectory(config.home)
